[allocator] Generalize recording to a pool
Previously the allocator would query whether a stream was recording a graph,
and look up the pool associated with a graph. This change has the allocator
directly associate a stream with a mempool, decoupling "record this stream to a pool"
from the action of "record all actions to a cuda graph".

ghstack-source-id: c84ec4d23935458a87c59e28a323d0429af08ce9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/96542